requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
